requirements/common.txt
python-telegram-bot[all]==21.1.1
SQLAlchemy==2.0.29